Struct aws_sdk_datasync::client::fluent_builders::CreateLocationS3

pub struct CreateLocationS3<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateLocationS3<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateLocationS3Output, SdkError<CreateLocationS3Error>> where
R::Policy: SmithyRetryPolicy<CreateLocationS3InputOperationOutputAlias, CreateLocationS3Output, CreateLocationS3Error, CreateLocationS3InputOperationRetryAlias>,

pub fn subdirectory(self, inp: impl Into<String>) -> Self

A subdirectory in the Amazon S3 bucket. This subdirectory in Amazon S3 is used to read data from the S3 source location or write data to the S3 destination.

pub fn set_subdirectory(self, input: Option<String>) -> Self

pub fn s3_bucket_arn(self, inp: impl Into<String>) -> Self

The ARN of the Amazon S3 bucket. If the bucket is on an Amazon Web Services Outpost, this must be an access point ARN.

pub fn set_s3_bucket_arn(self, input: Option<String>) -> Self

pub fn s3_storage_class(self, inp: S3StorageClass) -> Self

The Amazon S3 storage class that you want to store your files in when this location is used as a task destination. For buckets in Amazon Web Services Regions, the storage class defaults to Standard. For buckets on Outposts, the storage class defaults to Amazon Web Services S3 Outposts.

For more information about S3 storage classes, see Amazon S3 Storage Classes. Some storage classes have behaviors that can affect your S3 storage cost. For detailed information, see Considerations when working with S3 storage classes in DataSync.

pub fn set_s3_storage_class(self, input: Option<S3StorageClass>) -> Self

pub fn s3_config(self, inp: S3Config) -> Self

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role that is used to access an Amazon S3 bucket.

For detailed information about using such a role, see Creating a Location for Amazon S3 in the DataSync User Guide.

pub fn set_s3_config(self, input: Option<S3Config>) -> Self

pub fn agent_arns(self, inp: impl Into<String>) -> Self

Appends an item to AgentArns.

To override the contents of this collection use set_agent_arns.

If you are using DataSync on an Amazon Web Services Outpost, specify the Amazon Resource Names (ARNs) of the DataSync agents deployed on your Outpost. For more information about launching a DataSync agent on an Amazon Web Services Outpost, see Deploy your DataSync agent on Outposts.

pub fn set_agent_arns(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<TagListEntry>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

The key-value pair that represents the tag that you want to add to the location. The value can be an empty string. We recommend using tags to name your resources.

pub fn set_tags(self, input: Option<Vec<TagListEntry>>) -> Self