Module aws_sdk_datasync::error

Modules

internal_exception

See InternalException

invalid_request_exception

See InvalidRequestException

Structs

CancelTaskExecutionError

CreateAgentError

CreateLocationEfsError

CreateLocationFsxWindowsError

CreateLocationNfsError

CreateLocationObjectStorageError

CreateLocationS3Error

CreateLocationSmbError

CreateTaskError

DeleteAgentError

DeleteLocationError

DeleteTaskError

DescribeAgentError

DescribeLocationEfsError

DescribeLocationFsxWindowsError

DescribeLocationNfsError

DescribeLocationObjectStorageError

DescribeLocationS3Error

DescribeLocationSmbError

DescribeTaskError

DescribeTaskExecutionError

InternalException

This exception is thrown when an error occurs in the DataSync service.

InvalidRequestException

This exception is thrown when the client submits a malformed request.

ListAgentsError

ListLocationsError

ListTagsForResourceError

ListTaskExecutionsError

ListTasksError

StartTaskExecutionError

TagResourceError

UntagResourceError

UpdateAgentError

UpdateLocationNfsError

UpdateLocationObjectStorageError

UpdateLocationSmbError

UpdateTaskError

UpdateTaskExecutionError

Enums

CancelTaskExecutionErrorKind

CreateAgentErrorKind

CreateLocationEfsErrorKind

CreateLocationFsxWindowsErrorKind

CreateLocationNfsErrorKind

CreateLocationObjectStorageErrorKind

CreateLocationS3ErrorKind

CreateLocationSmbErrorKind

CreateTaskErrorKind

DeleteAgentErrorKind

DeleteLocationErrorKind

DeleteTaskErrorKind

DescribeAgentErrorKind

DescribeLocationEfsErrorKind

DescribeLocationFsxWindowsErrorKind

DescribeLocationNfsErrorKind

DescribeLocationObjectStorageErrorKind

DescribeLocationS3ErrorKind

DescribeLocationSmbErrorKind

DescribeTaskErrorKind

DescribeTaskExecutionErrorKind

ListAgentsErrorKind

ListLocationsErrorKind

ListTagsForResourceErrorKind

ListTaskExecutionsErrorKind

ListTasksErrorKind

StartTaskExecutionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAgentErrorKind

UpdateLocationNfsErrorKind

UpdateLocationObjectStorageErrorKind

UpdateLocationSmbErrorKind

UpdateTaskErrorKind

UpdateTaskExecutionErrorKind