Module aws_sdk_datasync::model

Modules

agent_list_entry

See AgentListEntry

ec2_config

See Ec2Config

filter_rule

See FilterRule

location_filter

See LocationFilter

location_list_entry

See LocationListEntry

nfs_mount_options

See NfsMountOptions

on_prem_config

See OnPremConfig

options

See Options

private_link_config

See PrivateLinkConfig

s3_config

See S3Config

smb_mount_options

See SmbMountOptions

tag_list_entry

See TagListEntry

task_execution_list_entry

See TaskExecutionListEntry

task_execution_result_detail

See TaskExecutionResultDetail

task_filter

See TaskFilter

task_list_entry

See TaskListEntry

task_schedule

See TaskSchedule

Structs

AgentListEntry

Represents a single entry in a list of agents. AgentListEntry returns an

Ec2Config

The subnet and the security group that DataSync uses to access target EFS file system.

FilterRule

Specifies which files, folders, and objects to include or exclude when transferring files

LocationFilter

You can use API filters to narrow down the list of resources returned by ListLocations.

LocationListEntry

Represents a single entry in a list of locations. LocationListEntry

NfsMountOptions

Represents the mount options that are available for DataSync to access an NFS

OnPremConfig

A list of Amazon Resource Names (ARNs) of agents to use for a Network File System (NFS)

Options

Represents the options that are available to control the behavior of a

PrivateLinkConfig

The VPC endpoint, subnet, and security group that an agent uses to access IP addresses in

S3Config

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) role

SmbMountOptions

Represents the mount options that are available for DataSync to access an SMB

TagListEntry

Represents a single entry in a list of AWS resource tags. TagListEntry

TaskExecutionListEntry

Represents a single entry in a list of task executions.

TaskExecutionResultDetail

Describes the detailed result of a TaskExecution operation. This result

TaskFilter

You can use API filters to narrow down the list of resources returned by ListTasks.

TaskListEntry

Represents a single entry in a list of tasks. TaskListEntry returns an

TaskSchedule

Specifies the schedule you want your task to use for repeated executions. For more

Enums

AgentStatus

Atime

EndpointType

FilterType

Gid

LocationFilterName

LogLevel

Mtime

NfsVersion

ObjectStorageServerProtocol

Operator

OverwriteMode

PhaseStatus

PosixPermissions

PreserveDeletedFiles

PreserveDevices

S3StorageClass

SmbSecurityDescriptorCopyFlags

SmbVersion

TaskExecutionStatus

TaskFilterName

TaskQueueing

TaskStatus

TransferMode

Uid

VerifyMode