Module aws_sdk_datasync::operation

Structs

CancelTaskExecution

Cancels execution of a task.

CreateAgent

Activates an DataSync agent that you have deployed on your host. The activation

CreateLocationEfs

Creates an endpoint for an Amazon EFS file system.

CreateLocationFsxWindows

Creates an endpoint for an Amazon FSx for Windows File Server file system.

CreateLocationNfs

Defines a file system on a Network File System (NFS) server that can be read from or

CreateLocationObjectStorage

Creates an endpoint for a self-managed object storage bucket. For more information

CreateLocationS3

Creates an endpoint for an Amazon S3 bucket.

CreateLocationSmb

Defines a file system on a Server Message Block (SMB) server that can be read from or

CreateTask

Creates a task.

DeleteAgent

Deletes an agent. To specify which agent to delete, use the Amazon Resource Name (ARN)

DeleteLocation

Deletes the configuration of a location used by DataSync.

DeleteTask

Deletes a task.

DescribeAgent

Returns metadata such as the name, the network interfaces, and the status (that is,

DescribeLocationEfs

Returns metadata, such as the path information about an Amazon EFS location.

DescribeLocationFsxWindows

Returns metadata, such as the path information about an Amazon FSx for Windows File Server

DescribeLocationNfs

Returns metadata, such as the path information, about an NFS location.

DescribeLocationObjectStorage

Returns metadata about a self-managed object storage server location. For more information

DescribeLocationS3

Returns metadata, such as bucket name, about an Amazon S3 bucket location.

DescribeLocationSmb

Returns metadata, such as the path and user information about an SMB location.

DescribeTask

Returns metadata about a task.

DescribeTaskExecution

Returns detailed metadata about a task that is being executed.

ListAgents

Returns a list of agents owned by an Amazon Web Services account in the Amazon Web Services Region specified in the

ListLocations

Returns a list of source and destination locations.

ListTagsForResource

Returns all the tags associated with a specified resource.

ListTaskExecutions

Returns a list of executed tasks.

ListTasks

Returns a list of all the tasks.

StartTaskExecution

Starts a specific invocation of a task. A TaskExecution value represents

TagResource

Applies a key-value pair to an Amazon Web Services resource.

UntagResource

Removes a tag from an Amazon Web Services resource.

UpdateAgent

Updates the name of an agent.

UpdateLocationNfs

Updates some of the parameters of a previously created location for Network File System (NFS) access.

UpdateLocationObjectStorage

Updates some of the parameters of a previously created location for self-managed object

UpdateLocationSmb

Updates some of the parameters of a previously created location for Server Message Block

UpdateTask

Updates the metadata associated with a task.

UpdateTaskExecution

Updates execution of a task.